backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.7
redis>=4.6.0
celery>=5.3.1
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
